Add xTOPI CSR support
xTOPI CSR provide information about the highest-priority
pending-and-enabled interrupt for machine/supervisor level.

This CSR serves as an basic support of SxAIA extension, as
the Linux kernel and openSBI only require this CSR fo SxAIA.
In addition, this CSR affects nothing and only provides
interrupt information, so enable it unconditionally.

Signed-off-by: Inochi Amaoto <inochiama@gmail.com>